rag-mkdocs/requirements.txt
# IMPORTANT: These versions have dependency conflicts!
# Use Poetry for installation: poetry install
# Poetry will automatically resolve compatible versions

# Compatible versions (installed via Poetry):
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-text-splitters>=0.2.0,<0.3.0
faiss-cpu==1.8.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
prometheus-client>=0.20.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
httpx>=0.27.0
email-validator>=2.1.0

